po-175jx: extract matchesTargetRole to shared config module
Move the triplicated matchesTargetRole() definition from ashby.ts,
greenhouse.ts, and lever.ts into config.ts alongside TARGET_ROLES.
All three source files now import the single implementation. Consolidated
the three identical matchesTargetRole test blocks into config.test.ts.
