+++ src/tests/job-hunter/greenhouse.test.ts
@@ -13,7 +13,6 @@
 import {
   fetchGreenhouseJobs,
   isRemote,
-  matchesTargetRole,
   normalizeJob,
 } from '../../job-hunter/sources/greenhouse';
 import type { GreenhouseJob } from '../../job-hunter/sources/greenhouse';
@@ -77,35 +76,6 @@ const remoteInContentJob: GreenhouseJob = {
 
 afterEach(() => {
   jest.restoreAllMocks();
-});
-
-// ─── matchesTargetRole() ──────────────────────────────────────────────────────
-
-describe('matchesTargetRole()', () => {
-  it('returns true for an exact target role title', () => {
-    expect(matchesTargetRole('VP of Engineering')).toBe(true);
-  });
-
-  it('returns true for a title that contains the target role', () => {
-    expect(matchesTargetRole('Senior VP of Engineering')).toBe(true);
-  });
-
-  it('returns false for an unrelated title', () => {
-    expect(matchesTargetRole('Software Engineer')).toBe(false);
-  });
-
-  it('matches target roles case-insensitively', () => {
-    expect(matchesTargetRole('vp of engineering')).toBe(true);
-  });
-
-  it('returns false for an empty title', () => {
-    expect(matchesTargetRole('')).toBe(false);
-  });
-
-  it('uses a provided roles list instead of the default', () => {
-    expect(matchesTargetRole('Custom Role', ['Custom Role'])).toBe(true);
-    expect(matchesTargetRole('VP of Engineering', ['Custom Role'])).toBe(false);
-  });
 });
 
 // ─── isRemote() ──────────────────────────────────────────────────────────────
